% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{ndiff}
\alias{ndiff}
\title{Normalized Difference tool}
\usage{
ndiff(x, y)
}
\arguments{
\item{x}{un objeto con valores numericos}

\item{y}{un objeto con valores numericos de la misma clase que x}
}
\value{
Uvalores numericos del indice
}
\description{
\code{ndiff} calcula el indice de diferencia normalizada
}
\examples{
ndiff(0.8,0.05)
ndiff(c(0.3,0.9),c(0.7,0.4))

}
