% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erkikevectorial.R
\name{sig_buffer}
\alias{sig_buffer}
\title{Buffer GIS tools}
\usage{
sig_buffer(capa_entrada, distancia, dissolve = TRUE, campo = NULL)
}
\arguments{
\item{capa_entrada}{un objeto de tipo vectorial de la clase sf}

\item{distancia}{Distancia de vecindad para determinar la zona de influencia en m}

\item{dissolve}{Parámetro para controlar la fusión de geometrias.}

\item{campo}{Parámetro para fusionar por un atributo. Opcional.}
}
\value{
Un objeto vectorial de la case sfc_POLYGON
}
\description{
\code{sig_buffer} devuelve la zona de influencia entorno a las geometrias de un shapefile
}
\examples{
\dontrun{
ciudades_data <- data.frame(
nombre =c("Madrid", "Barcelona", "Sevilla"),
poblacion = c(3, 2, 5),
lon=c(3, 2, 5),
lat=c(5, 3, 2)
)
ciudades_sf <- st_as_sf(ciudades_data, coords = c("lon","lat"), crs =4326)

sig_buffer(ciudades_sf,distancia=200)
sig_buffer(ciudades_sf, distancia=200, dissolve = TRUE)
sig_buffer(ciudades_sf, distancia=200, dissolve = FALSE)}
}
